refactor: merge RGT files, revert non-model changes to upstream, remove hardcoded backend
- Merge rgt_heads into rgt.py, move mappings to examples/rgt/
- Merge 4 manifold files into rgt_layers.py
- Revert mpops/torch.py, ops/*/__init__.py, download.py to upstream
- Remove hardcoded TL_BACKEND from all RGT files
- Set TL_BACKEND via run_server.py environment
