+++ examples/rgt/train.py
@@ -1,82 +1,5 @@
-import os
-os.environ['TL_BACKEND'] = 'torch'  # Set backend before importing TLX/GammaGL
-
 import tensorlayerx as tlx
 import torch
-import torch.optim._functional as _torch_opt_f
-import tensorlayerx.optimizers.torch_optimizers as _tlx_torch_opt
-
-# Set PyTorch configurations
-torch.set_default_dtype(torch.float32)
-if torch.cuda.is_available():
-    torch.backends.cudnn.benchmark = True
-
-# PyTorch 2.6 changed torch.load default weights_only=True.
-_orig_torch_load = torch.load
-
-def _compat_torch_load(*args, **kwargs):
-    if 'weights_only' not in kwargs:
-        kwargs['weights_only'] = False
-    return _orig_torch_load(*args, **kwargs)
-
-torch.load = _compat_torch_load
-
-# Compat for mixed torch internals
-def _compat_named_parameters(self, prefix='', recurse=True, remove_duplicate=True):
-    memo = set()
-    modules = self.named_modules(prefix=prefix) if recurse else [(prefix, self)]
-    for module_prefix, module in modules:
-        for name, param in module._parameters.items():
-            if param is None:
-                continue
-            if remove_duplicate and param in memo:
-                continue
-            if remove_duplicate:
-                memo.add(param)
-            full_name = f"{module_prefix}.{name}" if module_prefix else name
-            yield full_name, param
-
-torch.nn.Module.named_parameters = _compat_named_parameters
-
-# Compat for torch Adam functional signature drift across versions.
-_orig_f_adam = _torch_opt_f.adam
-
-def _normalize_state_steps(state_steps, params):
-    if state_steps is None:
-        return state_steps
-    device = None
-    if isinstance(params, (list, tuple)) and len(params) > 0 and hasattr(params[0], "device"):
-        device = params[0].device
-    normalized = []
-    for step in state_steps:
-        if torch.is_tensor(step):
-            if step.numel() == 1:
-                normalized.append(step.reshape(()))
-            else:
-                normalized.append(step)
-        else:
-            t = torch.tensor(step, device=device)
-            normalized.append(t.reshape(()))
-    return normalized
-
-def _compat_f_adam(*args, **kwargs):
-    kwargs.setdefault("maximize", False)
-    if "state_steps" in kwargs:
-        params = kwargs.get("params", args[0] if len(args) > 0 else None)
-        kwargs["state_steps"] = _normalize_state_steps(kwargs["state_steps"], params)
-    elif len(args) >= 6:
-        args = list(args)
-        args[5] = _normalize_state_steps(args[5], args[0])
-        args = tuple(args)
-    try:
-        return _orig_f_adam(*args, **kwargs)
-    except TypeError:
-        kwargs.pop("maximize", None)
-        return _orig_f_adam(*args, **kwargs)
-
-_torch_opt_f.adam = _compat_f_adam
-_tlx_torch_opt.F.adam = _compat_f_adam
-
 import numpy as np
 import os
 import random
